# api/openapi.yaml
openapi: 3.1.0
info:
  title: Silt API
  version: 0.1.0
  description: |
    A self-hosted change journal for Docker Compose stacks.

    This document is hand-maintained: there is no OpenAPI generator for stdlib
    net/http handlers, and adding one is not worth a dependency at this size.
    It is the contract, not a description of one — `openapi-typescript`
    generates the frontend types from it, and a Go contract test asserts the
    handlers still match, so drift fails CI instead of surfacing in a browser.

    All timestamps are Unix milliseconds in UTC.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
...
paths:
...
  /api/projects:
    get:
      tags: [projects]
      summary: List Compose projects
      operationId: listProjects
      parameters:
        - name: host
          in: query
          description: Host id. Defaults to the only host on a single-host install.
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
  /api/projects/{id}:
    get:
      tags: [projects]
      summary: Get one project
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/projects/{id}/snapshots:
    get:
      tags: [snapshots]
      summary: List a project's snapshots, newest first
      operationId: listSnapshots
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: before
          in: query
          description: Only snapshots taken before this timestamp.
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: changed_only
          in: query
          description: Only snapshots where the configuration changed.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Snapshots
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Snapshot" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/projects/{id}/services:
    get:
      tags: [projects]
      summary: Service names seen in a project
      operationId: listProjectServices
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Service names
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/projects/{id}/services/{service}:
    get:
      tags: [projects]
      summary: One service's history
      description: |
        Image identity, state and restart counts over time, plus the points at
        which each environment key's value changed. Environment values are
        keyed digests unless the key is on the cleartext keep-list, so this
        answers "when did SECRET_KEY change?" without ever holding the secret.
      operationId: getServiceHistory
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: service
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 200, maximum: 1000 }
      responses:
        "200":
          description: Service history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceHistory" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/projects/{id}/snapshot:
    post:
      tags: [snapshots]
      summary: Take a snapshot now
      operationId: takeSnapshot
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "202":
          description: Snapshot taken
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502":
          description: The snapshot could not be taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: The collector is not running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/snapshots/{id}:
    get:
      tags: [snapshots]
      summary: Get one snapshot with its service states
      operationId: getSnapshot
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SnapshotDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/snapshots/{id}/compose:
    get:
      tags: [snapshots]
      summary: The effective, redacted project model for a snapshot
      description: |
        A record of what ran, with secret values already replaced by keyed
        digests. It is not a file to re-apply.
      operationId: getCompose
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: format
          in: query
          schema: { type: string, enum: [json, yaml], default: json }
      responses:
        "200":
          description: Project model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectModel" }
            application/yaml:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/snapshots/{id}/files:
    get:
      tags: [snapshots]
      summary: Compose files captured in a snapshot
      operationId: listSnapshotFiles
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Captured files
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ComposeFile" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/snapshots/{id}/file:
    get:
      tags: [snapshots]
      summary: The redacted text of one captured file
      description: |
        Line structure is preserved exactly; only values are replaced, so line
        numbers and surrounding context match the file as written.
      operationId: getSnapshotFile
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/diff/file:
    get:
      tags: [snapshots]
      summary: Line-by-line diff of one file between two snapshots
      description: |
        Answers "which line changed", which the structured diff cannot: that
        reports an environment key moved, not where in the file it lives.
        A file absent from one side diffs against empty, rendering a creation
        or deletion rather than an error.
      operationId: getFileDiff
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: to
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: path
          in: query
          required: true
          schema: { type: string }
        - name: context
          in: query
          description: Unchanged lines around each hunk, or `full` for the whole file.
          schema: { type: string, default: "3" }
      responses:
        "200":
          description: File diff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileDiff" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/projects/{id}/files:
    get:
      tags: [projects]
      summary: Every file path this project has captured
      operationId: listProjectFilePaths
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: File paths
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/projects/{id}/files/preview:
    get:
      tags: [projects]
      summary: Live preview of a file for marking
      description: |
        Reads the file fresh from disk, applies the redaction a capture would
        apply, and stores nothing. The marking UI works from this because a
        stored capture is already redacted and would leave nothing to decide
        about. Only paths this project has referenced can be previewed.
      operationId: previewFile
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Preview lines
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FilePreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: No compose roots configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/projects/{id}/redaction-rules:
    get:
      tags: [projects]
      summary: Manual redaction decisions for a project
      operationId: listRedactionRules
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RedactionRule" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      tags: [projects]
      summary: Hide or reveal a value
      description: |
        Applied at capture time, before anything is written, so a hidden value
        is never stored rather than stored and later concealed. Revealing
        affects only future captures: earlier snapshots hold a digest, not the
        value. Adding a rule clears any opposing rule for the same target.
      operationId: createRedactionRule
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RedactionRuleRequest" }
      responses:
        "201":
          description: Rule stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedactionRule" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/projects/{id}/redaction-rules/{rule}:
    delete:
      tags: [projects]
      summary: Remove a manual redaction decision
      operationId: deleteRedactionRule
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: rule
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: Removed
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/diff:
    get:
      tags: [snapshots]
      summary: Structural diff between two snapshots
      operationId: getDiff
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: to
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Diff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Diff" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/events:
    get:
      tags: [events]
      summary: List events, newest first
      operationId: listEvents
      parameters:
        - name: from
          in: query
          schema: { type: integer, format: int64 }
        - name: to
          in: query
          schema: { type: integer, format: int64 }
        - name: project
          in: query
          schema: { type: integer, format: int64 }
        - name: service
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: severity
          in: query
          schema: { type: string, enum: [info, warn, error] }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
  /api/timeline:
    get:
      tags: [events]
      summary: Change markers and events merged onto one axis
      operationId: getTimeline
      parameters:
        - name: from
          in: query
          schema: { type: integer, format: int64 }
        - name: to
          in: query
          schema: { type: integer, format: int64 }
        - name: project
          in: query
          schema: { type: integer, format: int64 }
        - name: bucket
          in: query
          description: |
            Bucket width as a Go duration (`5m`, `1h`). Clamped server-side so
            no request can ask for an unbounded number of buckets.
          schema: { type: string }
      responses:
        "200":
          description: Timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Timeline" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/ingest:
    post:
      tags: [events]
      summary: Generic external event webhook
      description: |
        Accepts a loose body so Uptime Kuma, a cron `curl` and a Home Assistant
        automation all work without a custom integration. Authenticate with
        `Authorization: Bearer <token>` or `?token=`; not every webhook source
        can set headers. Returns 503 when `SILT_INGEST_TOKEN` is unset — an
        unconfigured endpoint is closed, not open.
      operationId: ingest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestRequest" }
      responses:
        "202":
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Body too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Ingest is not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/auth/login:
    get:
      tags: [system]
      summary: Start an OpenID Connect login
      description: |
        Redirects to the configured provider's authorization endpoint using the
        authorization code flow with PKCE. The per-login state, nonce and code
        verifier ride back in a short-lived, path-scoped cookie; nothing is
        stored server-side, because there is exactly one consumer and it
        arrives with the callback.

        `next` is where to send the browser afterwards. It is reduced to a
        same-origin path before it is used, so this cannot become an open
        redirect.
      operationId: startOIDCLogin
      parameters:
        - name: next
          in: query
          schema: { type: string }
          description: Same-origin path to return to. Anything else becomes "/".
      responses:
        "302":
          description: Redirect to the provider
        "503":
          description: OpenID Connect is not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/callback:
    get:
      tags: [system]
      summary: Finish an OpenID Connect login
      description: |
        Verifies the state, exchanges the code with the PKCE verifier,
        validates the id_token and its nonce, applies the group and user
        allowlists, and starts a session.

        Failures redirect back to the app with a `login_error` query
        parameter rather than returning JSON: the browser arrived here by
        navigation, and a JSON error would be a dead end with no way back.
      operationId: finishOIDCLogin
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: error
          in: query
          schema: { type: string }
          description: Set by the provider when it refused the login.
      responses:
        "302":
          description: Redirect to the app, signed in or carrying login_error
        "503":
          description: OpenID Connect is not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/setup:
    post:
      tags: [system]
      summary: Claim the built-in account with its first password
      description: |
        Public, because on a fresh install there is no way to authenticate yet,
        and gated on the account still being unclaimed, so it works exactly
        once. Every other endpoint is refused until it succeeds, which is what
        keeps the setup screen from being decoration.

        Setting `SILT_PASSWORD_HASH` claims the account before Silt starts, and
        removes this window entirely.
      operationId: setupAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: The password was set and a session started
        "400":
          description: The password does not meet the policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Already set up
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/password:
    put:
      tags: [system]
      summary: Change the built-in account's password
      description: |
        The current password is required even though the caller is signed in: a
        session someone walked away from should not be enough to lock its owner
        out. Every other session is revoked on success, so changing the
        password because you think it leaked also ends whatever leaked.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current, password]
              properties:
                current: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Changed, and this session restarted
        "400":
          description: The current password is wrong, or the new one is too weak
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not signed in as the built-in account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/account:
    put:
      tags: [system]
      summary: Turn the built-in account on or off
      description: |
        Refused when it would leave no way to sign in — configure a provider or
        a reverse proxy first.
      operationId: setAccountEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: The new state
        "409":
          description: Turning it off would lock everyone out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: The built-in account is disabled by configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/search:
    get:
      tags: [read]
      summary: Search projects, services, environment keys, files and events
      description: |
        A host running forty stacks has no other way to answer "when did
        anything about radarr change?" — the alternative is opening projects one
        at a time until you find it.

        A substring match, not a wildcard pattern: the terms people type contain
        `_` and `%`, and there is nothing to escape. A query shorter than two
        characters returns nothing, because one character matches most of the
        database and answers nothing.

        Environment *values* are never searched. They are keyed digests, and the
        cleartext ones are on the keep-list precisely because they are not
        secrets worth hunting for.

        Each category is capped independently, so a term matching a thousand
        events still shows the one project it matched.
      operationId: search
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: The term. Shorter than two characters returns nothing.
      responses:
        "200":
          description: What matched, by category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResults" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Project:
      type: object
      required: [id, host_id, name, config_files, first_seen_at, last_seen_at, archived]
      properties:
        id: { type: integer, format: int64 }
        host_id: { type: integer, format: int64 }
        name: { type: string }
        working_dir: { type: string }
        config_files:
          type: array
          items: { type: string }
        first_seen_at: { type: integer, format: int64 }
        last_seen_at: { type: integer, format: int64 }
        archived: { type: boolean }
...